refactor(anthropic): inline the key policy into sanitize-keys
Fold `key-policy.ts` into `sanitize-keys.ts` now that the real test suite runs
(the split existed mainly to keep the policy runtime-pure for isolated probing).
The Anthropic constraint is a private `KeySanitizationPolicy` next to the bound
`sanitizeSchemaPropertyKeys`, so the provider's key handling is one cohesive
module with a single import surface. No behavior change.
